feat: implement query-based search with text result display
- Add getLineFromFile() helper to efficiently read specific lines from wikitext.txt
  by line number (avoids loading entire file into memory)
- Add searchWithQueryFile() function to:
  * Load collection from disk
  * Read query vector from binary file (assumes float32 format)
  * Perform k-nearest neighbor search with configurable parameters
  * Display results with Wikipedia text snippets (75 chars preview)

- Command-line interface: ./arrowDB query <query_file> [k] [ef]
  Example: ./arrowDB query test_query.bin 10 200

- Result display format: [Score: 0.95] Text snippet...

This enables end-to-end semantic search workflow:
1. Generate query embedding from text
2. Load binary query file
3. Search collection
4. Display matching Wikipedia articles with similarity scores

+++ include/arrow/utils/utils.h
@@ -131,10 +131,10 @@ inline DistanceMetric jsonToDistanceMetric(const json& j) {
  */
 inline json dataTypeToJson(DataType dtype) {
     switch (dtype) {
-        case DataType::Int16:
-            return "Int16";
-        case DataType::Float16:
-            return "Float16";
+        case DataType::Int32:
+            return "Int32";
+        case DataType::Float32:
+            return "Float32";
         default:
             throw std::runtime_error("Unknown DataType");
     }
@@ -145,10 +145,10 @@ inline json dataTypeToJson(DataType dtype) {
  */
 inline DataType jsonToDataType(const json& j) {
     std::string dtypeStr = j.get<std::string>();
-    if (dtypeStr == "Int16") {
-        return DataType::Int16;
-    } else if (dtypeStr == "Float16") {
-        return DataType::Float16;
+    if (dtypeStr == "Int32") {
+        return DataType::Int32;
+    } else if (dtypeStr == "Float32") {
+        return DataType::Float32;
     } else {
         throw std::runtime_error("Unknown DataType: " + dtypeStr);
     }
